requirements.txt
terminusdb-client
pydantic
typing
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]